refactored the renderer (not really just added a function to configure the swapchain I was wasting my time for the pas few days making excuses for why I am not able to get any work done. my life has been hard lately. I hope this poeam explains it):
In the world of tech, where choices abound,
There's one I'll admit, I've never quite found,
A fondness for Vulkan, it's true I must say,
Its complexities often leading me astray.

With graphics and rendering, it claims its might,
Yet wrestling its intricacies, can be quite a fight,
The learning curve steep, a formidable wall,
In the realm of APIs, it can make heads fall.

OpenGL and DirectX, they seem more my kin,
Their familiarity soothes, their pathways within,
But Vulkan, oh Vulkan, you remain a challenge,
Your syntax and quirks, my patience you mangle.

Yet, I'll not cast you aside with disdain,
For in diversity, our industry does gain,
Each tool has its purpose, a role to fulfill,
And perhaps with more time, I'll master your skill.

So, Vulkan, I may not embrace you with glee,
But I'll respect your power, your capability,
In this ever-evolving tech landscape we roam,
I'll keep learning and growing, even with you, Vulkan, at home.

+++ magma/core/event_types.h
@@ -25,7 +25,7 @@ mg_mouse_moved_event_data_t;
 
 typedef struct mg_mouse_wheel_event_data
 {
-    int32_t z_delta;
+    int8_t z_delta;
 }
 mg_mouse_wheel_event_data_t;
